+++ panimg/contrib/oct_converter/readers/fda.py
@@ -1,3 +1,4 @@
+import logging
 from pathlib import Path
 
 import numpy as np
@@ -8,6 +9,8 @@ from panimg.contrib.oct_converter.image_types import (
     FundusImageWithMetaData,
     OCTVolumeWithMetaData,
 )
+
+logger = logging.getLogger(__name__)
 
 
 class FDA:
@@ -90,9 +93,9 @@ class FDA:
                     chunk_location = f.tell()
                     f.seek(chunk_size, 1)
                     chunk_dict[chunk_name] = [chunk_location, chunk_size]
-        print(f"File {self.filepath} contains the following chunks:")
+        logger.debug(f"File {self.filepath} contains the following chunks:")
         for key in chunk_dict.keys():
-            print(key)
+            logger.debug(key)
         return chunk_dict
 
     def read_oct_volume(self):